# apps/api/docs/swagger.yaml
basePath: /
definitions:
...
  graph.cleanupResponse:
    properties:
      deleted_entity_count:
        type: integer
      deleted_relationship_count:
        type: integer
      matched_entity_count:
        type: integer
      matched_relationship_count:
        type: integer
      workspace_id:
        type: string
      workspace_path:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Local-first pipeline API for ingesting and reasoning over engineering
    context.
  title: ContextOS API
  version: "1.0"
paths:
...
  /graph:
    get:
      description: Returns persisted canonical entities for a workspace, optionally
        filtered by entity type.
      parameters:
      - description: Workspace path or ID
        in: query
        name: workspace_id
        required: true
        type: string
      - description: Filter by entity type (e.g. feature, person, service)
        in: query
        name: entity_type
        type: string
      - description: Include low-signal regex entities and co-occurrence-only relationships
        in: query
        name: include_noise
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Query workspace entity graph
      tags:
      - graph
  /graph/cleanup:
    post:
      description: Permanently removes backend-classified low-signal graph entity
        and relationship rows for a workspace without deleting source artifacts, findings,
        chat history, or connected sources.
      parameters:
      - description: Workspace path or ID
        in: query
        name: workspace_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/graph.cleanupResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Clean noisy graph data
      tags:
      - graph
...
swagger: "2.0"